pyproject.toml
[project]
name = "example-gcp-batch"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Ren Adachi", email = "49895682+moeyashi@users.noreply.github.com" }
]
dependencies = [
    "google-cloud-batch>=0.16.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true
